+++ CHANGELOG.md
@@ -333,3 +333,4 @@
 * Define experimental `||=` method to apply a function inside transformer
 * Define `subview`, `subplace` and `substitute` methods in `Substructure` typeclass
 * Put `sub` method into `Substructure` typeclass
+* Remove `^.`, `.~` and `%~` infix optics methods